tsfx::feature_extractors

Module extras

Enums

ChunkAggregator

Functions

_absolute_energy 🔒

_absolute_maximum 🔒

_absolute_sum_of_changes 🔒

_agg_linear_trend 🔒

_aggregate_on_chunks 🔒

_autocorrelation 🔒

_c3 🔒

_cid_ce 🔒

_count_above 🔒

_count_above_mean 🔒

_count_below 🔒

_count_below_mean 🔒

_first_location_of_maximum 🔒

_first_location_of_minimum 🔒

_get_length_sequences_where 🔒

_has_duplicate 🔒

_has_duplicate_max 🔒

_has_duplicate_min 🔒

_index_mass_quantile 🔒

_kurtosis 🔒

_large_standard_deviation 🔒

_last_location_of_maximum 🔒

_last_location_of_minimum 🔒

_linear_trend 🔒

_longest_strike_above_mean 🔒

_longest_strike_below_mean 🔒

_make_nan_struct_column 🔒

_make_nan_struct_column_int 🔒

_mean_absolute_change 🔒

_mean_change 🔒

_mean_n_absolute_max 🔒

_median_mut 🔒

Return the median. Sorts its argument in place.

_number_crossing_m 🔒

_number_peaks 🔒

_out 🔒

_percentage_of_reoccurring_values_to_all_datapoints 🔒

_percentage_of_reoccurring_values_to_all_values 🔒

_quantile 🔒

_range_count 🔒

_ratio_beyond_r_sigma 🔒

_ratio_value_number_to_time_series_length 🔒

_roll 🔒

_sum_of_reoccurring_data_points 🔒

_sum_of_reoccurring_values 🔒

_symmetry_looking 🔒

_time_reversal_asymmetry_statistic 🔒

_variation_coefficient 🔒

absolute_energy

Abolute energy feature.

absolute_maximum

absolute_sum_of_changes

agg_linear_trend 🔒

autocorrelation

c3

calculate_sequential_ols

Calculates the Ordinary Least Squares (OLS) slope and intercept for a sequence of y-values where x-values are assumed to be a sequential index (0, 1, 2, …, N-1). Returns a tuple of (intercept, slope).

cid_ce

count_above

count_above_mean

count_below

count_below_mean

expr_abs_energy

The absolute energy of the time series cf. absolute_energy, calculated using the native Polars API.

expr_kurtosis

Kurtosis implemented using the native Polars API. See kurtosis.

expr_mean_change

Mean change implemented using the native Polars API. See mean_change.

expr_quantile

extra_aggregators

first_location_of_maximum

first_location_of_minimum

has_duplicate

has_duplicate_max

has_duplicate_min

index_mass_quantile

kurtosis

Kurtosis feature.

large_standard_deviation

last_location_of_maximum

last_location_of_minimum

linear_trend

longest_strike_above_mean

longest_strike_below_mean

mean_absolute_change

Mean absolute change feature.

mean_change

The mean change of a time series is defined as $$ \text{mean change} = \frac{1}{n-1} \sum_{i=1}^{n-1} x_{i + 1} - x_{i} = \frac{x_{n} - x_1}{n-1} $$.

mean_n_absolute_max

number_crossing_m

number_peaks

percentage_of_reoccurring_values_to_all_datapoints

percentage_of_reoccurring_values_to_all_values

range_count

ratio_beyond_r_sigma

ratio_value_number_to_time_series_length

sum_of_reoccurring_data_points

sum_of_reoccurring_values

symmetry_looking

time_reversal_asymmetry_statistic

variance_larger_than_standard_deviation

variation_coefficient