tsfx::feature_extractors

Module minimal

Functions

_max 🔒

_mean 🔒

_min 🔒

_out 🔒

_rms 🔒

_skewness 🔒

_standard_deviation 🔒

_sum_values 🔒

_variance 🔒

count

Length feature.

expr_maximum

The maximum value of the time series, using the native Polars API

expr_mean

The mean of all values of the time series, using the native Polars API

expr_median

Median feature.

expr_minimum

The minimum value of the time series, using the native Polars API

expr_root_mean_square

The root mean square of all values of the time series, using the native Polars API

expr_skewness

The skewness of all values in the time series, using the native Polars API

expr_standard_deviation

The standard deviation of all values of the time series, using the native Polars API

expr_sum

The sum of all values of the time series, using the native Polars API

expr_variance

The variance of all values of the time series, using the native Polars API

maximum

Maximum feature.

mean

Mean feature.

minimal_aggregators

minimum

Minimum feature.

root_mean_square

Root mean square feature.

skewness

Skewness feature.

standard_deviation

Standard deviation feature.

sum_values

Sum of values feature.

variance

Variance feature.